# internal/zstd: make val inlineable