@AlexanderYastrebov 2023-09-26
See https://go-review.googlesource.com/c/go/+/530717#message-455ce4779561da2d1e39982c65437e25d18e9323 discussion